LayoutTests/fast/js/isPrototypeOf-expected.txt

This tests that isPrototypeOf works according to the ECMA spec.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.isPrototypeOf('string') is false
PASS Array.prototype.isPrototypeOf(new Array()) is true
PASS ObjectA.prototype.isPrototypeOf(objb) is true
PASS ObjectA.prototype.isPrototypeOf(objc) is true
PASS successfullyParsed is true

TEST COMPLETE
